Refactor zipexe into its own package, and fixed it to close the darn file !
