+++ rice/append.go
@@ -11,6 +11,8 @@ import (
 	"runtime"
 	"strings"
 	"time"
+
+	"github.com/GeertJohan/go.rice/zipexe"
 )
 
 func operationAppend(pkg *build.Package) {
@@ -59,10 +61,12 @@ func operationAppend(pkg *build.Package) {
 	verbosef("Will append to file: %s\n", binfileName)
 
 	// check that command doesn't already have zip appended
-	if rd, _ := zipExeOpen(binfileName); rd != nil {
+	rd, _, err := zipexe.Open(binfileName)
+	if err != nil {
 		fmt.Printf("Cannot append to already appended executable. Please remove %s and build a fresh one.\n", binfileName)
 		os.Exit(1)
 	}
+	rd.Close()
 
 	// open binfile
 	binfile, err := os.OpenFile(binfileName, os.O_WRONLY, os.ModeAppend)
